+++ packages/std/src/coin.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{errors::CoinFromStrError, math::Uint128};
 
-#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "std", derive(JsonSchema))]
 pub struct Coin {
     pub denom: String,
@@ -19,6 +19,12 @@ impl Coin {
             amount: Uint128::new(amount),
             denom: denom.into(),
         }
+    }
+}
+
+impl fmt::Debug for Coin {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "Coin {{ {} \"{}\" }}", self.amount, self.denom)
     }
 }
 
@@ -237,5 +243,11 @@ mod tests {
                 .to_string(),
             "Invalid amount: number too large to fit in target type"
         );
+    }
+
+    #[test]
+    fn debug_coin() {
+        let coin = Coin::new(123, "ucosm");
+        assert_eq!(format!("{:?}", coin), r#"Coin { 123 "ucosm" }"#);
     }
 }